Implement TS control protocol for Diag CAN.
TS can be either controlled by the VCU in the vehicle or with the Diag software (i.e. during charging). To ensure that the TS is not enabled unintentionally by the user of the software while connected to the vehicle, the control has to be requested first. In this case, the TS can only be controlled by the software. If the control is not requested, the TS can only be controlled by the VCU. Not fully tested yet, second PCAN needed.